Websockets and testcases made
