+++ tests/utils/image.png
