--- tests/ws/success.test.js
@@ -1,5 +1,0 @@
-const utils = require("../utils/testUtils.js");
-
-describe("Unit tests", () => {
-    
-});
